Create LICENSE
This project is licensed under the MIT License, which means you're free to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the code, as long as you include the original license and copyright notice. The software is provided "as is", without warranty of any kind, and the authors are not liable for any issues that arise from its use.